+++ esp-hal/src/psram/mod.rs
@@ -1,38 +1,35 @@
-#![cfg_attr(docsrs, procmacros::doc_replace(
-    "octal" => {
-        cfg(psram_octal_spi) => "Either `Octal` or `Quad` PSRAM will be used, depending on the setting of `ESP_HAL_CONFIG_PSRAM_MODE`.",
-        _ => ""
-    }
-))]
+#![cfg_attr(docsrs, procmacros::doc_replace)]
 //! # PSRAM (Pseudo-static RAM, SPI RAM) driver
 //!
 //! ## Overview
 //!
-//! The `PSRAM` module provides support for accessing and controlling the
-//! `Pseudo Static Random Access Memory (PSRAM)`. `PSRAM` provides additional
-//! external memory to supplement the internal memory of the MCU, allowing for
-//! increased storage capacity and improved performance in certain applications.
-//!
-//! The mapped start address for PSRAM depends on the amount of mapped flash
-//! memory.
-//!
+//! This module provides support to interface with `PSRAM` devices connected to the MCU.
+//! PSRAM provides additional external memory to supplement the internal memory of the MCU,
+//! allowing for increased storage capacity and improved performance in certain applications.
+#![doc = ""]
+#![cfg_attr(
+    psram_octal_spi,
+    doc = concat!("The ", chip_pretty!(), " can use either Quad SPI or Octal SPI to interface with PSRAM.
+        You need to configure the correct interface type using `ESP_HAL_CONFIG_PSRAM_MODE`.")
+)]
+#![doc = ""]
 //! ## Examples
 //!
 //! ### PSRAM as heap memory
 //!
 //! This example shows how to use PSRAM as heap-memory via esp-alloc.
-//! You need an MCU with at least 2 MB of PSRAM memory.
-//! # {octal}
 //!
-//! > The PSRAM example **must** be built in release mode!
+//! <section class="warning">
+//! The PSRAM example <em>must</em> be built in release mode!
+//! </section>
 //!
 //! ```rust, ignore
 //! # {before_snippet}
-//! # extern crate alloc;
-//! # use alloc::{string::String, vec::Vec};
-//! #
+//! extern crate alloc;
+//! use alloc::{string::String, vec::Vec};
+//!
 //! // Add PSRAM to the heap.
-//! esp_alloc::psram_allocator!(&peripherals.PSRAM, esp_hal::psram);
+//! esp_alloc::psram_allocator!(peripherals.PSRAM, esp_hal::psram);
 //!
 //! let mut large_vec: Vec<u32> = Vec::with_capacity(500 * 1024 / 4);
 //!
@@ -46,17 +43,16 @@
 
 use core::ops::Range;
 
-#[cfg(feature = "psram")]
-#[cfg_attr(docsrs, doc(cfg(feature = "psram")))]
 #[cfg_attr(esp32, path = "esp32.rs")]
 #[cfg_attr(esp32s2, path = "esp32s2.rs")]
 #[cfg_attr(esp32s3, path = "esp32s3.rs")]
 #[cfg_attr(any(esp32c5, esp32c61), path = "esp32c5_c61.rs")]
 pub(crate) mod implem;
 
-#[cfg(feature = "psram")]
-#[cfg_attr(docsrs, doc(cfg(feature = "psram")))]
 pub use implem::*;
+use portable_atomic::{AtomicUsize, Ordering};
+
+use crate::peripherals::PSRAM;
 
 /// Size of PSRAM
 ///
@@ -73,7 +69,6 @@ pub enum PsramSize {
 }
 
 impl PsramSize {
-    #[cfg_attr(not(feature = "psram"), expect(unused))]
     pub(crate) fn get(&self) -> usize {
         match self {
             PsramSize::AutoDetect => 0,
@@ -81,40 +76,43 @@ impl PsramSize {
         }
     }
 
-    #[cfg_attr(not(feature = "psram"), expect(unused))]
     pub(crate) fn is_auto(&self) -> bool {
         matches!(self, PsramSize::AutoDetect)
     }
 }
 
-/// Returns the address and size of the available in external memory.
-#[cfg(feature = "psram")]
-pub fn psram_raw_parts(_psram: &crate::peripherals::PSRAM<'_>) -> (*mut u8, usize) {
-    let range = psram_range();
-    (range.start as *mut u8, range.end - range.start)
-}
-
-// Using static mut should be fine since we are only writing to it once during
-// initialization. As other tasks and interrupts are not running yet, the worst
-// that can happen is, that the user creates a DMA buffer before initializing
-// the HAL. This will access the PSRAM range, returning an empty range - which
-// is, at that point, true. The user has no (safe) means to allocate in PSRAM
-// before initializing the HAL.
-#[cfg(feature = "psram")]
-static mut MAPPED_PSRAM: MappedPsram = MappedPsram { memory_range: 0..0 };
+static MAPPED_PSRAM_START: AtomicUsize = AtomicUsize::new(0);
+static MAPPED_PSRAM_END: AtomicUsize = AtomicUsize::new(0);
 
 pub(crate) fn psram_range() -> Range<usize> {
-    cfg_if::cfg_if! {
-        if #[cfg(feature = "psram")] {
-            #[allow(static_mut_refs)]
-            unsafe { MAPPED_PSRAM.memory_range.clone() }
-        } else {
-            0..0
-        }
-    }
+    let end = MAPPED_PSRAM_END.load(Ordering::Acquire);
+    let start = MAPPED_PSRAM_START.load(Ordering::Relaxed);
+    if end < start { 0..0 } else { start..end }
 }
 
-#[cfg(feature = "psram")]
-pub(crate) struct MappedPsram {
-    memory_range: Range<usize>,
+/// # Safety
+///
+/// This function must only be called once.
+unsafe fn set_psram_range(range: Range<usize>) {
+    MAPPED_PSRAM_START.store(range.start, Ordering::Relaxed);
+    MAPPED_PSRAM_END.store(range.end, Ordering::Release);
+}
+
+/// Enables externally-connected Pseudo-static RAM.
+pub struct Psram {
+    _peri: PSRAM<'static>,
+}
+
+impl Psram {
+    /// Initializes PSRAM.
+    pub fn new(peri: PSRAM<'static>, config: PsramConfig) -> Self {
+        init_psram(config);
+        Self { _peri: peri }
+    }
+
+    /// Returns the address and size of the available in external memory.
+    pub fn raw_parts(&self) -> (*mut u8, usize) {
+        let range = psram_range();
+        (range.start as *mut u8, range.end - range.start)
+    }
 }