+++ esp-hal/CHANGELOG.md
@@ -148,6 +148,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `esp_hal::interrupt::map` (#4996, #5007)
 - `InterruptHandler::new_not_nested` (#5000)
 - `esp_hal::interrupt::Priority::None` (#4996)
+- The `psram` feature has been replaced by a `Psram` driver object. esp-alloc has been updated so that no user code changes should be necessary when using `esp_alloc::psram_allocator!` (#5310)
+- The `esp_hal::psram::psram_raw_parts` function has been removed. Use `esp_alloc::psram_allocator`, or `Psram::raw_parts`. (#5310)
 
 ## [v1.0.0] - 2025-10-30